add scene 3d
